# Problem: tx format conversions not necessary