--- a/app/ante/setup.go
+++ b/app/ante/setup.go
@@ -107,30 +107,25 @@ func ValidateEthBasic(ctx sdk.Context, tx sdk.Tx, evmParams *evmtypes.Params, ba
 
 		txGasLimit += msgEthTx.GetGas()
 
-		txData, err := evmtypes.UnpackTxData(msgEthTx.Data)
-		if err != nil {
-			return errorsmod.Wrap(err, "failed to unpack MsgEthereumTx Data")
-		}
-
+		tx := msgEthTx.AsTransaction()
 		// return error if contract creation or call are disabled through governance
-		if !enableCreate && txData.GetTo() == nil {
+		if !enableCreate && tx.To() == nil {
 			return errorsmod.Wrap(evmtypes.ErrCreateDisabled, "failed to create new contract")
-		} else if !enableCall && txData.GetTo() != nil {
+		} else if !enableCall && tx.To() != nil {
 			return errorsmod.Wrap(evmtypes.ErrCallDisabled, "failed to call contract")
 		}
 
-		if baseFee == nil && txData.TxType() == ethtypes.DynamicFeeTxType {
+		if baseFee == nil && tx.Type() == ethtypes.DynamicFeeTxType {
 			return errorsmod.Wrap(ethtypes.ErrTxTypeNotSupported, "dynamic fee tx not supported")
 		}
 
-		ethTx := ethtypes.NewTx(txData.AsEthereumData())
-		if !allowUnprotectedTxs && !ethTx.Protected() {
+		if !allowUnprotectedTxs && !tx.Protected() {
 			return errorsmod.Wrapf(
 				errortypes.ErrNotSupported,
 				"rejected unprotected Ethereum transaction. Please EIP155 sign your transaction to protect it against replay-attacks")
 		}
 
-		txFee = txFee.Add(sdk.Coin{Denom: evmDenom, Amount: sdkmath.NewIntFromBigInt(txData.Fee())})
+		txFee = txFee.Add(sdk.Coin{Denom: evmDenom, Amount: sdkmath.NewIntFromBigInt(msgEthTx.GetFee())})
 	}
 
 	if !authInfo.Fee.Amount.Equal(txFee) {