--- a/x/evm/keeper/keeper.go
+++ b/x/evm/keeper/keeper.go
@@ -191,7 +191,7 @@ func (k *Keeper) SetHooks(eh types.EvmHooks) *Keeper {
 }
 
 // PostTxProcessing delegate the call to the hooks. If no hook has been registered, this function returns with a `nil` error
-func (k *Keeper) PostTxProcessing(ctx sdk.Context, msg core.Message, receipt *ethtypes.Receipt) error {
+func (k *Keeper) PostTxProcessing(ctx sdk.Context, msg *core.Message, receipt *ethtypes.Receipt) error {
 	if k.hooks == nil {
 		return nil
 	}
@@ -199,7 +199,7 @@ func (k *Keeper) PostTxProcessing(ctx sdk.Context, msg core.Message, receipt *et
 }
 
 // Tracer return a default vm.Tracer based on current keeper state
-func (k Keeper) Tracer(msg core.Message, rules params.Rules) vm.EVMLogger {
+func (k Keeper) Tracer(msg *core.Message, rules params.Rules) vm.EVMLogger {
 	return types.NewTracer(k.tracer, msg, rules)
 }