--- a/x/evm/types/tracer.go
+++ b/x/evm/types/tracer.go
@@ -36,7 +36,7 @@ const (
 
 // NewTracer creates a new Logger tracer to collect execution traces from an
 // EVM transaction.
-func NewTracer(tracer string, msg core.Message, rules params.Rules) vm.EVMLogger {
+func NewTracer(tracer string, msg *core.Message, rules params.Rules) vm.EVMLogger {
 	// TODO: enable additional log configuration
 	logCfg := &logger.Config{
 		Debug: true,